Add overlay for Ormolu.
Ormolu is available in nixpkgs-unstable, but this is not up-to-date for Darwin
at the time of writing (meaning _lots_ of compilation, including of GHC), so a
better choice on macOS is nixpkgs-19.09-darwin. This does not have Ormolu, hence
the overlay.

We choose a slightly outdated version of Ormolu because newer versions trigger
the same compile-the-world behaviour as we see when using nixpkgs-unstable.